Do not recursively require things to prevent a segfault, do not use file paths with extensions
When loading the various checksum files, on Ubuntu with Ruby 2.1.0 the checksum calculators are globbed from checksum.rb. Each checksum calculator, in turn, requires checksum.rb - which causes an infinite require loop when used with ActiveSupport (that is not smart enough to figure out those requires are repeating).

To avoid this problem, it is much better not to require the checksum.rb because it is already loaded once the globbing for the checksum files begins.

We also got problems from requiring paths ending with .rb on our particular Ruby/ActiveSupport combination.